divand.jl documentation

divand

divand.jl documentation

divand.divandrun
divand.divandgo
divand.divand_averaged_bg
divand.load_mask
divand.domain
divand.SDNMetadata
divand.save
divand.saveobs
divand.loadobs
divand.loadbigfile
divand.checkobs
divand.fit_isotropic
divand.fit
divand.smoothfilter
divand.Anam.loglin
divand.Anam.logit
divand.divadoxml
divand.random
divand.distance
divand.metric

Examples

To run the example, you need to install PyPlot. In the folder examples of divand, you can run e.g. the example divand_simple_example_1D.jl by issuing:

# cd("/path/to/divand/examples")
include("divand_simple_example_1D.jl")

Replace /path/to/divand/ by the installation directory of divand which is the output of Pkg.dir("divand") if you installed divand using Julias package manager.

Vocabulary

urn_str

divand.Vocab.CFVocab — Type.

collection = Vocab.CFVocab(url)

Create a Dict-like object represeting the NetCDF CF Standard Name vocabulary. If the url is not provided then current CF Standard Name list http://cfconventions.org/Data/cf-standard-names/current/src/cf-standard-name-table.xml is used. Individual standard names are retirved by indexing which return an object of the type CFEntry:

collection = Vocab.CFVocab()
entry = collection["sea_water_temperature"]

source

Base.haskey — Method.

bool = haskey(collection::CFVocab,stdname)

Return true if stdname is part of the NetCDF CF Standard Name vocabulary collection.

source

divand.Vocab.SDNCollection — Function.

collection = SDNCollection(name)

Open the SeaDataNet collection with the name name at the URL http://www.seadatanet.org/urnurl/collection/ The collection can be indexed with brackets using the identifier.

using divand
collection = Vocab.SDNCollection("P01")
concept = collection["PSALPR01"]
@show Vocab.prefLabel(concept)

source

divand.Vocab.prefLabel — Function.

s = Vocab.prefLabel(c::Vocab.Concept)

Return the preferred label of a concept c

source

s = Vocab.prefLabel(urn::AbstractString)

Return the preferred label of a concept usings it URN (Uniform Resource Name)

source

divand.Vocab.altLabel — Function.

s = Vocab.altLabel(c::Vocab.Concept)

Return the alternative label of a concept c

source

s = Vocab.altLabel(urn::AbstractString)

Return the alternative label of a concept usings it URN (Uniform Resource Name)

source

divand.Vocab.notation — Function.

s = Vocab.notation(c::Vocab.Concept)

Return the identifier of a concept c

source

s = Vocab.notation(urn::AbstractString)

Return the identifier of a concept usings it URN (Uniform Resource Name)

source

Base.find — Method.

find(c::Concept,name,collection)

Return a list of related concepts in the collection collection. name can be the string "related", "narrower", "broader".

source

divand.Vocab.description — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

divand.Vocab.canonical_units — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

divand.Vocab.splitURL — Function.

collection,tag,key = Vocab.splitURL(url)

Split a concept URL into collection, tag and key. url must finishe with a slash.

source

Information for developers

Update the documentation

Install

Pkg.add("Documenter")

Troubleshooting

No plot windows

If the following command doesn't produce any figure

using PyPlot
plot(0, 1)

a possible solution is to modify the backend: this is done by editing the python configuration file matplotlibrc. The location of this file is obtained in python with:

import matplotlib
matplotlib.matplotlib_fname

which, in my case, returns '~/.config/matplotlib/matplotlibrc'