divand.jl documentation

divand

divand.jl documentation

divand.divandrun — Function.

divandrun(mask,pmn,xi,x,f,len,epsilon2; <keyword arguments>)

Perform an n-dimensional variational analysis of the observations f located at the coordinates x. The array fi represent the interpolated field at the grid defined by the coordinates xi and the scales factors pmn.

Input:

mask: binary mask delimiting the domain. true is inside and false outside. For oceanographic application, this is the land-sea mask where sea is true and land is false.

pmn: scale factor of the grid. pmn is a tuple with n elements. Every element represents the scale factor of the corresponding dimension. Its inverse is the local resolution of the grid in a particular dimension.

xi: tuple with n elements. Every element represents a coordinate of the final grid on which the observations are interpolated

x: tuple with n elements. Every element represents a coordinate of the observations

f: value of the observations minus the background estimate (m-by-1 array). (see note)

len: correlation length

epsilon2: error variance of the observations (normalized by the error variance of the background field). epsilon2 can be a scalar (all observations have the same error variance and their errors are decorrelated), a vector (all observations can have a difference error variance and their errors are decorrelated) or a matrix (all observations can have a difference error variance and their errors can be correlated). If epsilon2 is a scalar, it is thus the inverse of the signal-to-noise ratio.

Optional input arguments specified as keyword arguments

velocity: velocity of advection constraint. The default is no-advection constraint

alpha: alpha is vector of coefficients multiplying various terms in the cost function. The first element multiplies the norm. The other i-th element of alpha multiplies the (i+1)-th derivative. Per default, the highest derivative is m = ceil(1+neff/2) where neff is the effective dimension of the problem (the number of dimensions with a nonzero correlation length).

 The values of alpha is the (m+1)th row of the Pascal triangle:
    m=0         1
    m=1       1   1
    m=1     1   2   1     (n=1,2)
    m=2   1   3   3   1   (n=3,4)
    ...

constraints: a structure with user specified constrain

moddim: modulo for cyclic dimension (vector with n elements). Zero is used for non-cyclic dimensions. Halo points should not be included for cyclic dimensions. For example if the first dimension is cyclic, then the grid point corresponding to mask(1,j) should be between mask(end,1) (left neighbor) and mask(2,j) (right neighbor)

fracindex: fractional indices (n-by-m array). If this array is specified, then x and xi are not used.

inversion: direct solver (:chol for Cholesky factorization) or a interative solver (:pcg for preconditioned conjugate gradient) can be used.

compPC: function that returns a preconditioner for the primal formulation if inversion is set to 'pcg'. The function has the following arguments:

     fun = compPC(iB,H,R)

where iB is the inverse background error covariance, H the observation operator and R the error covariance of the observation. The function compPC returns the preconditioner fun(x,fx) computing fx = M x (the inverse of M times x) where M is a positive defined symmetric matrix [1]. Effectively, the system E⁻¹ A (E⁻¹)ᵀ (E x) = E⁻¹ b is solved for (E x) where E Eᵀ = M. Ideally, M should this be similar to A, so that E⁻¹ A (E⁻¹)ᵀ is close to the identity matrix.

fi0: starting field for iterative primal algorithm (same size as mask)

f0: starting field for iterative dual algorithm (same size as the observations f)

operatortype: Val{:sparse} for using sparse matrices (default) or Val{:MatFun} or using functions to define the constrains.

scale_len: true (default) if the correlation length-scale should be scaled such that the analysical kernel reaches 0.6019072301972346 (besselk(1.,1.)) at the same distance. The kernel behaves thus similar to the default kernel in two dimensions (alpha = [1,2,1]).

alphabc : numerical value defining how the last grid points are stretched outward. 1, the default value mimics an infinite domain.

To have previous behaviour of finite domain use alphabc=0

btrunc : if provided defines where to truncate the calculation of the covariance matrix B. Only values up and including alpha[btrunc] will be calculated. IF the

iterative solution is calculated, the missing terms will be calculated on the fly during the conjugate gradient calulcations. Default value is none and full covariance calculation.

Output:

fi: the analysed field

s: structure with an array s.P representing the analysed error covariance

Note:

If zero is not a valid first guess for your variable (as it is the case for e.g. ocean temperature), you have to subtract the first guess from the observations before calling divand and then add the first guess back in.

Example:

see divand_simple_example.jl

References

[1] https://en.wikipedia.org/w/index.php?title=Conjugate_gradient_method&oldid=761287292#The_preconditioned_conjugate_gradient_method

source

divand.divandgo — Function.

Compute a variational analysis of arbitrarily located observations.

fi,s = divandgo(mask,pmn,xi,x,f,len,epsilon2,errormethod; ...);

Perform an n-dimensional variational analysis of the observations f located at the coordinates x. The array fi represent the interpolated field at the grid defined by the coordinates xi and the scales factors pmn.

Input:

As for divandrun but as a higher level routine which will automatically create windowing etc it also include the definition of the errormethod

errormethod : :cpme (clever poormans method), :none or :exact

Output:

fi: the analysed field

s: structure with an array s.P representing the analysed error covariance

source

divand.divand_averaged_bg — Function.

fma,faanom = divand_averaged_bg(mask,pmn,xi,x,f,len,epsilon2,toaverage;moddim=[])

Input:

As for divandrun, including all dimensions before averaging

additional argument:

toaverage: Array of ndims of boolean telling if in the corresponding direction averaging must be done

Presently NO optional arguments from divandrun supported except moddim

Output:

fma: Analysis where in the directions where toaverage is true, the same value is found

faanom: Data anomalies when the analysis is subtracted from the input field.

source

divand.load_mask — Function.

deprecated

source

deprecated

source

xi,yi,mask = load_mask(bath_name,isglobal,xi,yi,level::Number)

Generate a land-sea mask based on the topography from the NetCDF file bathname. The parameter isglobal is true if the NetCDF file covers the whole globe and thus the last longitude point can be considered to be right next to the first longitude point. xi and yi is a vector of the longitude and latitude grid onto which the bathymetry should be interpolated. In the water, level is postive and in the air level is negative.

source

divand.domain — Function.

mask,(pm,pn),(xi,yi) = domain(bathname,bathisglobal,lonr,latr)

Generate a 2D geospatial domain based on the topography from the NetCDF file bathname.

source

mask,(pm,pn,po),(xi,yi,zi) = domain(bathname,bathisglobal,lonr,latr,depthr)

Generate a 3D geospatial domain based on the topography from the NetCDF file bathname. if zlevel = :surface, then depthr is zero for the sea surface and positive in water (positive is down) if zlevel = :floor, then depthr is zero for the sea floor and positive in water (positive is up)

source

mask,(pm,pn,po,pp),(xi,yi,zi,ti) = domain(bathname,bathisglobal,lonr,latr,depthr,timer)

Generate a geospatial domain based on the topography from the NetCDF file bathname.

source

divand.SDNMetadata — Function.

ncglobalattrib,ncvarattrib = SDNMetadata(metadata,fi)

Based on the information in the dictionary metadata and the analysed 4D field fi produce a list of NetCDF global and variable attributes for divand_save2.

source

divand.save — Function.

save(filename,xyi,fi,varname;
                      ncvarattrib = Dict(), ncglobalattrib = Dict(), ...)

Save the result of the analysis in a NetCDF file .

Input arguments

filename: the name of the NetCDF file

xyi: tuple with n vectors. Every element in this tuple represents a coordinate of the final grid on which the observations are interpolated

fi: the analysed field

varname: the name of the NetCDF variable

Optional arguments:

ncglobalattrib: a dictionary with the global attributes

ncvarattrib: a dictionary with the variable attributes

relerr: relative error

source

divand.checkobs — Function.

 checkobs(x,v,ids)
 checkobs(io::IO,x,v,ids)

Print some basic information about the coordinates x (tuple of vector) and values v (vector) having the identifier ids (vector of strings) to check erroneous data. It prints wheter NaNs or Infs are found and the minimum and maximum value.

If the argument io is provided, the information is input/output stream io.

source

divand.fit_isotropic — Function.

var0,len,distx,covar,fitcovar = fit_isotropic(x,v,distbin,min_count;
                           alpha = divand.alpha_default(length(x)),
                           len = 1.,
                           var0 = 1.,
                           minlen = 0.,
                           maxlen = 10.,
                           minvar0 = 0.,
                           maxvar0 = 10.,
                           tolrel = 1e-4,
                           maxpoints = 1000000,
                           distfun = (xi,xj) -> sqrt(sum(abs2,xi-xj))),
                           progress = (var,len,fitness) -> nothing
                       )

Determines the optimal correlation length len and variance (for a separation distance approaching zero) var0 of a cloud of data points with value v and coordiantes x (tuple of vectors with the coordinates).

The function can find the solution corresponding to a local minimum which is not necessarily the global minimum.

See also empiriccovar for future information about the output parameters.

Optional input parameters:

alpha: if one correlation length is forced to zero during the anaylsis the values of alpha sould be set using the effective dimension. For example, if a 2D-analysis is simulated by forcing the vertical correlation length to zero, then alpha should be set to [1,2,1], otherwise alpha will be [1,3,3,1] (for for any proper 3D analysis).

len: initial value for the correlation length

var0: initial value of the variance

minlen, maxlen: minmum and maximum value for the correlation length

minvar0, maxvar0: minmum and maximum value for the variance

tolrel: relative tolerance for the solver

maxpoints: maximum number of data points considered

distfun: function to compute the distance between point xi (vector) and xj. Per default distun is the Eucedian distance (xi,xj) -> sqrt(sum(abs2,xi-xj))).

progress: call-back function to show the progress of the optimization with the input parameters var, len and fitness (all scalars).

source

divand.fit — Function.

var0opt,lensopt,distx,covar,fitcovar = fit(x,v,distbin,min_count;
         alpha = divand.alpha_default(length(x)),
         minlen = zeros(length(x)),
         maxlen = ones(length(x)),
         tolrel = 1e-4,
         lens0 = ones(length(x)),
         var0 = 1.,
         minvar0 = 0.,
         maxvar0 = 2.,
         maxpoints = 1000000,
         distfun = (xi,xj,lens) -> sqrt(sum(abs2,(xi-xj)./lens)),
         progress = (iter,var,len,fitness) -> nothing
         )

The same as the function fit_isotropic except that now the correlation length-scale lens0, minlen, maxlen, lensopt are a vectors (one value per dimension). The distance function distfun uses an additional parameter to compute the normalized distance.

See the note of alpha in divafit which also applies here.

source

Examples

To run the example, you need to install PyPlot. In the folder examples of divand, you can run e.g. the example divand_simple_example_1D.jl by issuing:

# cd("/path/to/divand/examples")
include("divand_simple_example_1D.jl")

Replace /path/to/divand/ by the installation directory of divand which is the output of Pkg.dir("divand") if you installed divand using Julias package manager.

Vocabulary

urn_str

divand.Vocab.CFVocab — Type.

collection = Vocab.CFVocab(url)

Create a Dict-like object represeting the NetCDF CF Standard Name vocabulary. If the url is not provided then current CF Standard Name list http://cfconventions.org/Data/cf-standard-names/current/src/cf-standard-name-table.xml is used. Individual standard names are retirved by indexing which return an object of the type CFEntry:

collection = Vocab.CFVocab()
entry = collection["sea_water_temperature"]

source

Base.haskey — Method.

bool = haskey(collection::CFVocab,stdname)

Return true if stdname is part of the NetCDF CF Standard Name vocabulary collection.

source

divand.Vocab.SDNCollection — Function.

collection = SDNCollection(name)

Open the SeaDataNet collection with the name name at the URL http://www.seadatanet.org/urnurl/collection/ The collection can be indexed with brackets using the identifier.

using divand
collection = Vocab.SDNCollection("P01")
concept = collection["PSALPR01"]
@show Vocab.prefLabel(concept)

source

divand.Vocab.prefLabel — Function.

s = Vocab.prefLabel(c::Vocab.Concept)

Return the preferred label of a concept c

source

s = Vocab.prefLabel(urn::AbstractString)

Return the preferred label of a concept usings it URN (Uniform Resource Name)

source

divand.Vocab.altLabel — Function.

s = Vocab.altLabel(c::Vocab.Concept)

Return the alternative label of a concept c

source

s = Vocab.altLabel(urn::AbstractString)

Return the alternative label of a concept usings it URN (Uniform Resource Name)

source

divand.Vocab.notation — Function.

s = Vocab.notation(c::Vocab.Concept)

Return the identifier of a concept c

source

s = Vocab.notation(urn::AbstractString)

Return the identifier of a concept usings it URN (Uniform Resource Name)

source

Base.find — Method.

find(c::Concept,name,collection)

Return a list of related concepts in the collection collection. name can be the string "related", "narrower", "broader".

source

divand.Vocab.description — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

divand.Vocab.canonical_units — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

divand.Vocab.splitURL — Function.

collection,tag,key = Vocab.splitURL(url)

Split a concept URL into collection, tag and key. url must finishe with a slash.

source

Information for developpers

Update the documentation

Install

Pkg.add("Documenter")