DIVAnd.jl documentation

DIVAnd

DIVAnd.jl documentation

DIVAnd.diva3d — Function.

residuals = diva3d(xi,x,value,len,epsilon2,filename,varname)

Create a 3D analysis (or a series of 3D analyses) with DIVAnd using the observations value (vector) at the locations x (tuple of vectors) onto the regular grid defined by the vectors xi using the scaled observational error variance epsilon2 and the correlation length len. The result will be saved in the NetCDF file filename under the variable varname.

Inputs

xi: tuple with n elements. Every element represents a coordinate of the final grid on which the observations are interpolated

x: tuple with n elements. Every element represents a coordinate of the observations

value: value of the observations

len: tuple with n elements. Every element represents the correlation length. If fitcorrlen is true, then len can be the empty tuple () or a tuple containing 3 arrays of normalized correlation lengths which will be multiplied by the horizontal and vertical correlation lengths.

epsilon2: error variance of the observations (normalized by the error variance of the background field). epsilon2 can be a scalar (all observations have the same error variance and their errors are decorrelated), a vector (all observations can have a different error variance and their errors are decorrelated) or a matrix (all observations can have a different error variance and their errors can be correlated). If epsilon2 is a scalar, it is thus the inverse of the signal-to-noise ratio.

filename: The output NetCDF filename.

varname: The name of the variable (used in the NetCDF file).

Optional input arguments:

bathname: path to the NetCDF bathymetry (default ../../DIVAnd-example-data/Global/Bathymetry/gebco_30sec_16.nc relative to this source file)

bathisglobal: true (default) is the bathymetry is a global data set

plotres: Call-back routine for plotting ((timeindex,sel,fit,erri) -> nothing)

timeorigin: Time origin (default DateTime(1900,1,1,0,0,0))

moddim: modulo for cyclic dimension (vector with n elements). Zero is used for non-cyclic dimensions. Halo points should not be included for cyclic dimensions. For example if the first dimension is cyclic, then the grid point corresponding to mask[1,j] should be between mask[end,1] (left neighbor) and mask[2,j] (right neighbor). The default is [0,0,0],

zlevel: :surface (default) for surface analysis and :floor for analysis from the bottom floor.

ncvarattrib: dictionary of NetCDF variable attributes.

ncglobalattrib: dictionary of NetCDF global attributes.

transform: Anamorphosis transformation function (default: Anam.notransform()).

fitcorrlen: true of the correlation length is determined from the observation (default false).

fithorz_param: dictionary with additional optional parameters for fithorzlen.

fitvert_param: dictionary with additional optional parameters for fitvertlen.

distfun: function to compute the distance (default (xi,xj) -> DIVAnd.distance(xi[2],xi[1],xj[2],xj[1])).

mask: if different from nothing, then this mask overrides land-sea mask based on the bathymetry

(default nothing).

background: if different from nothing, then this parameter allows one

to load the background from a call-back function (default nothing).

background_espilon2_factor: multiplication for epsilon2 when computing the background (default 10.).

memtofit: keyword controlling how to cut the domain depending on the memory remaining available for inversion. It is not total memory (default 3).

niter_e: Number of iterations to estimate the optimal scale factor of epsilon2 using Desroziers et al. 2005 (doi: 10.1256/qj.05.108). The default is 1 (i.e. no optimization is done).

Any additional keywoard arguments understood by DIVAndgo can also be used here (e.g. velocity constrain)

source

DIVAnd.DIVAndrun — Function.

DIVAndrun(mask,pmn,xi,x,f,len,epsilon2; <keyword arguments>)

Perform an n-dimensional variational analysis of the observations f located at the coordinates x. The array fi represent the interpolated field at the grid defined by the coordinates xi and the scales factors pmn.

Input:

mask: binary mask delimiting the domain. true is inside and false outside.

For oceanographic application, this is the land-sea mask where sea is true and land is false.

pmn: scale factor of the grid. pmn is a tuple with n elements. Every element represents the scale factor of the corresponding dimension. Its inverse is the local resolution of the grid in a particular dimension. For example, in two dimensions, pmn is a tuple (pm,pn) where pm is the inverse of the local resolution in first dimension and pn is the the inverse of the local resolution in second dimension.

xi: tuple with n elements. Every element represents a coordinate of the final grid on which the observations are interpolated.

x: tuple with n elements. Every element represents a coordinate of the observations.

f: value of the observations minus the background estimate (vector of m elements where m is the number of observations). See also note.

len: tuple with n elements. Every element represents the correlation length for a given dimension.

epsilon2: error variance of the observations (normalized by the error variance of the background field). epsilon2 can be a scalar (all observations have the same error variance and their errors are decorrelated), a vector (all observations can have a different error variance and their errors are decorrelated) or a matrix (all observations can have a different error variance and their errors can be correlated). If epsilon2 is a scalar, it is thus the inverse of the signal-to-noise ratio.

Optional input arguments specified as keyword arguments

velocity: velocity of advection constraint. The default is no-advection constraint

alpha: alpha is vector of coefficients multiplying various terms in the cost function. The first element multiplies the norm. The other i-th element of alpha multiplies the (i+1)-th derivative. Per default, the highest derivative is m = ceil(1+neff/2) where neff is the effective dimension of the problem (the number of dimensions with a nonzero correlation length) and ceil is the ceiling function (rounding up).

   The values of alpha is the (m+1)th row of the Pascal triangle:
      m=0         1
      m=1       1   1
      m=1     1   2   1     (n=1,2)
      m=2   1   3   3   1   (n=3,4)
      ...

constraints: a structure with user specified constraints (see DIVAnd_addc).

moddim: modulo for cyclic dimension (vector with n elements). Zero is used for non-cyclic dimensions. One should not include a boundary zone (sometimes called a ghost zone or halo) for cyclic dimensions. For example if the first dimension is cyclic, then the grid point corresponding to mask[1,j] should be between mask[end,1] (left neighbor) and mask[2,j] (right neighbor).

fracindex: fractional indices (n-by-m array). If this array is specified, then x and xi are not used.

inversion: direct solver (:chol for Cholesky factorization) or an interative solver (:pcg for preconditioned conjugate gradient [1]) can be used.

compPC: function that returns a preconditioner for the primal formulation if inversion is set to 'pcg'. The function has the following arguments:

     fun = compPC(iB,H,R)

where iB is the inverse background error covariance, H the observation operator and R the error covariance of the observation. The function compPC returns the preconditioner fun(x,fx) computing fx = M \ x (the inverse of M times x) where M is a positive defined symmetric matrix [1]. Effectively, the system E⁻¹ A (E⁻¹)ᵀ (E x) = E⁻¹ b is solved for (E x) where E Eᵀ = M. Ideally, M should this be similar to A, so that E⁻¹ A (E⁻¹)ᵀ is close to the identity matrix.

fi0: starting field for iterative primal algorithm (same size as mask).

f0: starting field for iterative dual algorithm (same size as the observations f).

operatortype: Val{:sparse} for using sparse matrices (default) or Val{:MatFun} or using functions to define the constrains.

scale_len: true (default) if the correlation length-scale should be scaled such that the analytical kernel reaches 0.6019072301972346 (besselk(1.,1.)) at the same distance than in 2D. The kernel behaves thus similar to the default kernel in two dimensions (alpha = [1,2,1]).

alphabc : numerical value defining how the last grid points are stretched outward. If alphabc is 1, the default value mimics an infinite domain. To have previous behaviour of finite domain use alphabc equal to 0.

btrunc : if provided defines where to truncate the calculation of the covariance matrix B. Only values up and including alpha[btrunc] will be calculated. If the iterative solution is calculated, the missing terms will be calculated on the fly during the conjugate gradient calulcations. Default value is none and full covariance calculation.

Output:

fi: the analysed field

s: structure with an array s.P representing the analysed error covariance

Note:

If zero is not a valid first guess for your variable (as it is the case for e.g. ocean temperature), you have to subtract the first guess from the observations before calling DIVAnd and then add the first guess back in.

Example:

see DIVAnd_simple_example.jl

References

[1] https://en.wikipedia.org/w/index.php?title=Conjugate_gradient_method&oldid=761287292#The_preconditioned_conjugate_gradient_method

source

DIVAnd.DIVAndgo — Function.

fi, erri, residuals, qcvalues, scalefactore = DIVAndgo(mask,pmn,xi,x,f,len,epsilon2,errormethod; ...);

Input:

Same arguments as DIVAndrun with in addition

errormethod : you have the choice between :cpme (clever poorman's method, default method if parameter not provided), :none or :exact (only available if windowed analysis are done with DIVAndrun)

MEMTOFIT=: keyword controlling how to cut the domain depending on the memory remaining available for inversion (not total memory)

RTIMESONESCALES= : if you provide a tuple of length scales, data are weighted differently depending on the numbers of neighbours they have. See weight_RtimesOne for details

QCMETHOD= : if you provide a qc method parameter, quality flags are calculated. See DIVAnd_cv for details

Output:

fi: the analysed field

erri: relative error field on the same grid as fi. () if errormethod is fixed to :none

residuals: array of residuals at data points. For points not on the grid or on land: NaN

qcvalues: if QCMETHOD= is provided, the output array contains the quality flags otherwise qcvalues is (). For points on land or not on the grid: 0

scalefactore: Desroziers et al. 2005 (doi: 10.1256/qj.05.108) scale factor for epsilon2

Perform an n-dimensional variational analysis of the observations f located at the coordinates x. The array fi represent the interpolated field at the grid defined by the coordinates xi and the scales factors pmn.

IMPORTANT: DIVAndgo is very similar to DIVAndrun and is only interesting to use if DIVAndrun cannot fit into memory or if you want to parallelize. (In the latter case do not forget to define the number of workers; see addprocs for example)

source

DIVAnd.DIVAnd_averaged_bg — Function.

fma,faanom = DIVAnd_averaged_bg(mask,pmn,xi,x,f,len,epsilon2,toaverage;moddim=[])

Input:

As for DIVAndrun, including all dimensions before averaging

additional argument:

toaverage: Array of ndims of boolean telling if in the corresponding direction averaging must be done

Presently NO optional arguments from DIVAndrun supported except moddim

Output:

fma: Analysis where in the directions where toaverage is true, the same value is found

faanom: Data anomalies when the analysis is subtracted from the input field.

source

DIVAnd.SDNMetadata — Function.

ncglobalattrib,ncvarattrib = SDNMetadata(metadata,fi)

Based on the information in the dictionary metadata and the analysed 4D field fi produce a list of NetCDF global and variable attributes for DIVAnd_save2.

source

DIVAnd.save — Function.

save(filename,xyi,fi,varname;
                      ncvarattrib = Dict(), ncglobalattrib = Dict(), ...)

Save the result of the analysis in a NetCDF file .

Input arguments

filename: the name of the NetCDF file

xyi: tuple with n vectors. Every element in this tuple represents a coordinate of the final grid on which the observations are interpolated

fi: the analysed field

varname: the name of the NetCDF variable

Optional arguments:

ncglobalattrib: a dictionary with the global attributes

ncvarattrib: a dictionary with the variable attributes

relerr: relative error

source

DIVAnd.loadbigfile — Function.

value,lon,lat,depth,time,obsid = loadbigfile(filename)

Load data from the text file filename and returns vectors with the value, longitude, latitude, depth and time (as DateTime). A list string identifiers is also returned.

source

DIVAnd.checkobs — Function.

 checkobs(x,v,ids)
 checkobs(io::IO,x,v,ids)

Print some basic information about the coordinates x (tuple of vector) and values v (vector) having the identifier ids (vector of strings) to check erroneous data. It prints wheter NaNs or Infs are found and the minimum and maximum value.

If the argument io is provided, the information is input/output stream io.

source

DIVAnd.smoothfilter — Function.

ff = smoothfilter(x,f,scale)

Smooth the function f defined on x by solving the diffusion equation

∂ₜ ϕ = ν ∂²ₓ ϕ

scale is the spatial scales of the removed length-scales. It is defined as 2Tν where T is the integration time.

It uses the Greens functions for 1D diffusion: 1/sqrt(4 π ν t) * exp(-x^2 / (4νt))

source

DIVAnd.Anam.loglin — Function.

trans,invtrans = loglin(t; epsilon = 0.)

Provide the following transform log(x + epsilon) (for x < t) and its inverse. Beyond the threshold t (x ≥ t), the function is extended linearly in a continous way.

trans,invtrans are scalar functions such that for any x (x > epsilon), x == invtrans(trans(x)).

For any array X, we have: X == invtrans.(trans.(X)).

source

DIVAnd.Anam.logit — Function.

trans,invtrans = logit(; min = 0., max = 1.)

Provide the logit transform and its inverse. Per default the logit transform maps values within the interval from 0 and 1. This can be changed with the min and max parameters. Note that trans(min) = -∞ and trans(max) = +∞. The use safety-margin might be necessary.

source

DIVAnd.divadoxml — Function.

DIVAnd.divadoxml(filepath,varname,project,cdilist,xmlfilename;
                 ignore_errors = false)

Generate the XML metadata file xmlfilename from the NetCDF file filepath with the NetCDF variable varname. Project is either "SeaDataNet", "EMODNET-chemistry" or "SeaDataCloud". cdilist is the file from http://emodnet-chemistry.maris2.nl/download/export.zip.

The XML file contain a list of the data the originators. divadoxml will abort with an error if some combinations of EDMO code, local CDI ID are not present in the cdilist. Such errors can be ignore if ignore_errors is set to true.

source

DIVAnd.random — Function.

field = DIVAnd.random(mask,pmn,len,Nens)

Create Nens random fields with the correlation length len in a domain with the mask mask and the metric pmn.

See DIVAnd.DIVAndrun for more information about these parameters.

source

DIVAnd.distance — Function.

d = distance(lat1,lon1,lat2,lon2)

Compute the great-circle distance between the points (lat1,lon1) and (lat2,lon2). The units of all input and output parameters are degrees.

source

d = distance([lon1,lat1],[lon2,lat2])

The same as distance(lat1,lon1,lat2,lon2) but there the arguments are vectors and the order is longitude then latitude.

The units of all input and output parameters are degrees.

source

DIVAnd.interp — Function.

f = interp(xi,fi,x)

Interpolate field fi (n-dimensional array) defined at xi (tuble of n-dimensional arrays or vectors) onto grid x (tuble of n-dimensional arrays). The grid in xi must be align with the axis (e.g. produced by DIVAnd.ndgrid).

source

DIVAnd.backgroundfile — Function.

fun = backgroundfile(fname,varname)

Return a function fun which is used in DIVAnd to make anomalies out of observations based relative to the field defined in the NetCDF variable varname in the NetCDF file fname. It is assumed that the NetCDF variables has the variable lon, lat and depth. And that the NetCDF variable is defined on the same grid as the analysis.

source

DIVAnd.Quadtrees.checkduplicates — Function.

dupl = checkduplicates(x,value,delta,deltavalue)

Based the coordinates x (a tuple of longitude lons, latitudes lats, depth (zs) and time (times vector of DateTime)) check of points who are in the same spatio-temporal bounding box of a length delta. delta is a vector with 4 elements corresponding to longitude, latitude, depth and time (in days). dupl a vector of vectors containing indices of the duplicates.

source

dupl = checkduplicates(x1,value1,x2,v2,value2,delta,deltavalue)

Report duplicate of observation in data set (x2,v2) which are also in data set (x1,v1). x1 and x2 is a tuple of vectors with the cooridantes and v1 and v2 the corresponding values.

source

Bathymetry and spatial-temporal domain

DIVAnd.load_bath — Function.

xi,yi,bath = DIVAnd.load_bath(bath_name,isglobal,xi,yi)

Load the bathymetry from the NetCDF file bathname. The parameter isglobal is true if the NetCDF file covers the whole globe and thus the last longitude point can be considered to be right next to the first longitude point. xi and yi are vectors containing the longitude and latitude grid onto which the bathymetry should be interpolated.

source

DIVAnd.extract_bath — Function.

bx,by,b = DIVAnd.extract_bath(bath_name,isglobal,xi,yi)

Extract the bathymetry from the NetCDF file bathname. The parameter isglobal is true if the NetCDF file covers the whole globe and thus the last longitude point can be considered to be right next to the first longitude point. xi and yi are vectors defining the bounding box of the data. No interpolation is performed.

Convention: b is positive in the water and negative in the air.

source

DIVAnd.load_mask — Function.

xi,yi,mask = load_mask(bath_name,isglobal,xi,yi,level::Number)

Generate a land-sea mask based on the topography from the NetCDF file bathname. The parameter isglobal is true if the NetCDF file covers the whole globe and thus the last longitude point can be considered to be right next to the first longitude point. xi and yi are vectors containing the longitude and latitude grid onto which the bathymetry should be interpolated.

Convention: in the water, level is positive and in the air level is negative.

source

DIVAnd.DIVAnd_metric — Function.

pm,pn = DIVAnd_metric(lon,lat)

Compute metric scale factors pm and pn based on the arrays longitude lon and latitude lat. The variables pm and pn represent the inverse of the local resolution in meters using the mean Earth radius.

source

DIVAnd.domain — Function.

mask,(pm,pn),(xi,yi) = domain(bathname,bathisglobal,lonr,latr)

Generate a 2D geospatial domain based on the topography from the NetCDF file bathname.

source

mask,(pm,pn,po),(xi,yi,zi) = domain(bathname,bathisglobal,lonr,latr,depthr)

Generate a 3D geospatial domain based on the topography from the NetCDF file bathname. If zlevel is :surface, then depthr is zero for the sea surface and positive in water (positive is down). If zlevel is :floor, then depthr is zero for the sea floor and positive in water (positive is up)

source

mask,(pm,pn,po,pp),(xi,yi,zi,ti) = domain(bathname,bathisglobal,lonr,latr,depthr,timer)

Generate a geospatial domain based on the topography from the NetCDF file bathname.

source

DIVAnd.DIVAnd_rectdom — Function.

mask,pmn,xyi = DIVAnd_rectdom(coord1,coord2,...)

Create a "rectangular" domain in n dimensions with the coordinates coord1 coord2... assuming a Catersian metric. This functions returns the mask mask, the coordinates (xi,yi,...) and the metric (pm,pn...).

For example:

julia> mask,(pm,pn),(xi,yi) = DIVAnd_rectdom(linspace(0,1,50),linspace(0,1,50))

source

DIVAnd.DIVAnd_squaredom — Function.

mask,pmn,xyi = DIVAnd_squaredom(n,coord)

Create a "square" domain in n dimensions with the coordinates coord assuming a Catersian metric. This functions returns the mask mask, the coordinates (xi,yi,...) and the metric (pm,pn...).

Example

mask,(pm,pn),(xi,yi) = DIVAnd_squaredom(2,linspace(0,1,50))

source

DIVAnd.TimeSelectorYW — Function.

TS = TimeSelectorYW(years,yearwindow,monthlists)

The structure TS handles the time aggregation based on years and monthlists. It is similar to TimeSelectorYearListMonthList except that the elements of yearlists are centred around years and span yearwindow years. yearlists is in fact constructed by adding and subtracting yearwindow/2 to every element of years.

source

DIVAnd.TimeSelectorYearListMonthList — Type.

TS = TimeSelectorYearListMonthList(yearlists,monthlists)

The structure TS handles the time aggregation based on yearlists and monthlists. yearlists is a vector of ranges (containing start and end years), for example [1980:1990,1990:2000,2000:2010].

monthlists is a vector of two-element vector (containing start and end months), for example [1:3,4:6,7:9,10:12]

If a month range spans beyond December, then all Months must be specified, e.g. example [2:4,5:6,7:9,[10,11,12,1]] or [2:4,5:6,7:9,[10:12;1]]. However using [2:4,5:6,7:9,10:1] (bug!) will result in an empty month range.

Example

seasonal climatology using all data from 1900 to 2017

for winter (December-February), spring, summer, autumn

TS = DIVAnd.TimeSelectorYearListMonthList([1900:2017],[[12,1,2],[3,4,5],[6,7,8],[9,10,11]])

source

Load observations

DIVAnd.saveobs — Function.

DIVAnd.saveobs(filename,xy,ids;
               type_save = Float32,
               timeorigin = DateTime(1900,1,1,0,0,0),
               )

Save the location and time of the observation in the NetCDF file filename and their identifier ids. xy is a tuple with the vectors longitude, latitude, depth and time (as a vector of DateTime).

Optional arguments:

type_save: the type to save the data (default Float32). However, the time is always saved as Float64.

timeorigin: time origin for the time units attribute (default is

1900-01-01 00:00:00)

source

DIVAnd.saveobs(filename,varname,value,xy,ids;
               type_save = Float32,
               timeorigin = DateTime(1900,1,1,0,0,0),
               )

Save value and the location and time of the observation in the NetCDF file filename and their identifier ids. xy is a tuple with the vectors longitude, latitude, depth and time (as a vector of DateTime). The values will be saved in the variable called varname.

Optional arguments:

type_save: the type to save the data (default Float32). However, the time is always saved as Float64.

timeorigin: time origin for the time units attribute (default is

1900-01-01 00:00:00)

source

DIVAnd.loadobs — Function.

obsvalue,obslon,obslat,obsdepth,obstime,obsid = loadobs(T,filename,varname)

Load the variable varname from the NetCDF file filename. Coordinates (the NetCDF variables "obslon", "obslat", "obsdepth"), time ("obstime") and identifies ("obsids") will also be loaded. Numeric output arguments will have the type T.

source

DIVAnd.NCSDN.load — Function.

data,lon,lat,z,time,ids = load(T,fname::TS,param; qualityflags = [GOOD_VALUE, PROBABLY_GOOD_VALUE]) where TS <: AbstractString

source

data,lon,lat,z,time,ids = SDN.load(T,fnames,param; qualityflags = ...)

Load all data in the vector of file names fnames corresponding to the parameter param as the data type T. Only the data with the quality flags SDN.good_data and SDN.probably_good_data are loaded per default. The output parameters correspondata to the data, longitude, latitude, depth, time (as DateTime) and an identifier (as String).

source

DIVAnd.NCSDN.loadvar — Function.

data = loadvar(ds,param;
               fillvalue::T = NaN,
               qualityflags = [GOOD_VALUE, PROBABLY_GOOD_VALUE],
               qfname = param * QC_SUFFIX,
               )

Load the NetCDF variable param from the NCDataset ds. Data points not having the provide quality flags will be masked by fillvalue. qfname is the NetCDF variable name for the quality flags.

source

DIVAnd.ODVspreadsheet.loaddata — Function.

data = loaddata(sheet,profile,locname,fillvalue; fillmode = :repeat)

Load a single column referred by the local name locname in the profile profile from the ODV spreadsheet sheet. Empty values are either replaced by fillvalue (if fillmode is :fill) or the previous value if repeated (if fillmode is :repeat)

source

DIVAnd.ODVspreadsheet.parsejd — Function.

dt = parsejd(t)

Convert a Chronological Julian Day Number to a DateTime object. The reference value is taken from Chronological Julian Date

From the SDN standard: "A real number representing the Chronological Julian Date, which is defined as the time elapsed in days from 00:00 on January 1 st 4713 BC. ... "

The time origin is _not_ noon (12:00) on Monday, January 1, 4713 BC as for the Julia Date Number.

source

DIVAnd.ODVspreadsheet.myparse — Function.

v = myparse(T,s)

Parse the string s as a type T. Unlike Julia's parse function an error message contains the string s (which could not be parsed) for debugging.

source

Parameter optimization

DIVAnd.fit_isotropic — Function.

var0,len,distx,covar,fitcovar = fit_isotropic(x,v,distbin,mincount;
                           alpha = DIVAnd.alpha_default(length(x)),
                           minlen = 0.,
                           maxlen = 10.,
                           tolrel = 1e-4,
                           maxpoints = 10000,
                           nmean = 100,
                           distfun = (xi,xj) -> sqrt(sum(abs2,xi-xj))),
                           progress = (iter,var,len,fitness) -> nothing
                       )

Determines the optimal correlation length len and variance (for a separation distance approaching zero) var0 of a cloud of data points with value v and coordiantes x (tuple of vectors with the coordinates).

The function can find the solution corresponding to a local minimum which is not necessarily the global minimum.

See also empiriccovar for future information about the output parameters.

Optional input parameters:

alpha: if one correlation length is forced to zero during the anaylsis the values of alpha sould be set using the effective dimension. For example, if a 2D-analysis is simulated by forcing the vertical correlation length to zero, then alpha should be set to [1,2,1], otherwise alpha will be [1,3,3,1] (for any proper 3D analysis).

len: initial value for the correlation length.

minlen, maxlen: minimum and maximum values for the correlation length.

tolrel: relative tolerance for the optimizer.

maxpoints: maximum number of data points considered.

nmean: the number of times an empirical covariance is estimated. The average covariance is used for the fitting.

distfun: function to compute the distance between point xi (vector) and xj. Per default distfun is the Euclidian distance: (xi,xj) -> sqrt(sum(abs2,xi-xj))).

progress: call-back function to show the progress of the optimization with the input parameters iter, var, len and fitness (all scalars).

The length-scale parameters and the variance have the corresponding units from the x and v. It is therefore often necessary to provide reasonable values for these default parameters.

The algorithm used to estimate the correlation-length and variance is based on randomly choosen points. Therefore the result can be different if the function is invoked repeately. If nmean is increased, then these statistical fluctuations should decrease (for a not too large value of mincount, i.e. about 100 for most cases).

If the lower bound minlen is too small, then you might get the following error:

AmosException with id 4: input argument magnitude too large, complete loss of accuracy by argument reduction.

In these case, increase minlen.

source

DIVAnd.fit — Function.

var0opt,lensopt,distx,covar,fitcovar = fit(x,v,distbin,mincount;
         alpha = DIVAnd.alpha_default(length(x)),
         minlen = zeros(length(x)),
         maxlen = ones(length(x)),
         tolrel = 1e-4,
         lens0 = ones(length(x)),
         var0 = 1.,
         minvar0 = 0.,
         maxvar0 = 2.,
         maxpoints = 10000,
         distfun = (xi,xj,lens) -> sqrt(sum(abs2,(xi-xj)./lens)),
         progress = (iter,var,len,fitness) -> nothing
         )

The same as the function fit_isotropic except that now the correlation length-scale lens0, minlen, maxlen, lensopt are vectors (one value per dimension). The distance function distfun uses an additional parameter to compute the normalized distance.

The note of the optional parameters in divafit also applies here.

source

DIVAnd.DIVAnd_cv — Function.

bestfactorl,bestfactore, cvval,cvvalues, x2Ddata,y2Ddata,cvinter,xi2D,yi2D = DIVAnd_cv(mask,pmn,xi,x,f,len,epsilon2,nl,ne,method;...);

Input

Same as for DIVAndrun with three more parameters nl,ne and method

mask: binary mask delimiting the domain. true is inside and false outside. For oceanographic application, this is the land-sea mask.

pmn: scale factor of the grid. pmn is a tuple with n elements. Every element represents the scale factor of the corresponding dimension. Its inverse is the local resolution of the grid in a particular dimension.

xi: tuple with n elements. Every element represents a coordinate of the final grid on which the observations are interpolated

x: tuple with n elements. Every element represents a coordinate of the observations

f: value of the observations minus the background estimate (m-by-1 array). (see note)

len: correlation length

epsilon2: error variance of the observations (normalized by the error variance of the background field). epsilon2 can be a scalar (all observations have the same error variance and their errors are decorrelated), a vector (all observations can have a difference error variance and their errors are decorrelated) or a matrix (all observations can have a difference error variance and their errors can be correlated). If epsilon2 is a scalar, it is thus the inverse of the signal-to-noise ratio.

nl: number of testing points around the current value of L. 1 means one additional point on both sides of the current L. 0 is allowed and means the parameter is not optimised.

ne: number of testing points around the current value of epsilon2. 0 is allowed as for nl

method: cross validation estimator method 1: full CV 2: sampled CV 3: GCV 0: automatic choice between the three possible ones, default value

Optional input arguments specified via keyword arguments are the same as for DIVAnd

Output:

bestfactorl: best estimate of the multiplication factor to apply to len

bestfactore: best estimate of the multiplication factor to apply to epsilon2

cvvales : the cross validation values calculated

factors : the tested multiplication factors

cvinter : interpolated cv values for final optimisation

X2Data, Y2Data : coordinates of sampled cross validation in L,epsilon2 space . Normally only used for debugging or plotting

Xi2D, Yi2D : coordinates of interpolated estimator . Normally only used for debugging or plotting

The output bestfactorl and bestfactore represent multiplication factors which should be applied to L and epsilon2.

The len and epsilon2 provided should be close the real one as the tests will be performed around.

source

DIVAnd.empiriccovar — Function.

distx,covar,corr,varx,count = empiriccovar(x,v,distbin,mincount;
                          maxpoints = 10000,
                          distfun = (xi,xj) -> sqrt(sum(abs2,xi-xj)))

Compute the covariance, correlation and variance of a cloud of data points with the value v (a vector) and the location x (a tuple of vectors) grouped by distance. Random pairs are choosen and grouped by their distance (computed by distfun) in bins defined by distbin. The function try to fill at least mincount of data points in each bin but always stop after considering maxpoints pairs.

source

DIVAnd.fithorzlen — Function.

lenz,dbinfo = DIVAnd.fithorzlen(x,value,z)

Determines the horizontal correlation length lenz based on the measurments value at the location x (tuple of 3 vectors corresponding to longitude, latitude and depth).

Optional arguments:

smoothz (default 100): spatial filter for the correlation scale

searchz (default 50): vertical search distance

maxnsamp (default 5000): maximum number of samples

limitlen (default false): limit correlation length by mean distance between observations

source

DIVAnd.fitvertlen — Function.

lenz,dbinfo = DIVAnd.fitvertlen(x,value,z)

See also DIVAnd.fithorzlen

source

DIVAnd.lengraddepth — Function.

RL = lengraddepth(pmn,h, L;
                  h2 = h,
                  hmin = 0.001
                  )

Create the relative correlation length-scale field RL based on the bathymetry h and the metric pmn (tuple of arrays). Effectively the correlation-length scale is close to zero if the relative bathymetry gradients (|∇h|/h) are smaller than the length-scale L (in consistent units as pmn).

R_L = 1 / (1 + L |∇h| / max(h2,hmin))

Per default h2 is equal to h. The depth h must be positive. hmin must have the same units as h (usually meters).

source

DIVAnd.DIVAnd_cvestimator — Function.

theta = DIVAnd_cvestimator(s,residual)

Computes the cross validation estimator $(d-\hat{d})^T \mathbf R^{-1} (d-\hat{d}) / ( \mathbf 1^T \mathbf R^{-1} \mathbf 1)$ where the $\hat{d}$ is the analysis not using a data point.

source

DIVAnd.weight_RtimesOne — Function.

 weights = weight_RtimesOne(x,len)

Compute the weight of the observations at location x to reduce the influence of locally clustered data. x is a tuple with n elements. Every element represents a coordinate of the observations. len is a tuple of arrays representing the correlation length. len[i] is the correlation length in the i-th dimension.

source

DIVAnd.Rtimesx! — Function.

Rtimesx!(coord,LS,x,Rx)

Gaussian type R matix in ndim dimensions applied to vector x of length ndata. The Gaussian scale differs in each direction k : LS[k] Coordinates of point i are coord[i,1],coord[i,2],...,coord[i,ndim] To avoid an ndata² complexity a grid is set up first so as to allow only to calculate covarances when distances are smaller than 3*LS

Adapted from DIVA3D/src/Fortran/Util/Rtimesx_weighting.f90

source

Vocabulary

DIVAnd.Vocab.@urn_str — Macro.

urn"SDN:x:y:z'

Resolve a SeaDataNet URN (Uniform Resource Name) using https://www.seadatanet.org/urnurl/

source

DIVAnd.Vocab.CFVocab — Type.

collection = Vocab.CFVocab()
collection = Vocab.CFVocab(url = url)

Create a Dict-like object represeting the NetCDF CF Standard Name vocabulary. If the url is not provided then current CF Standard Name list http://cfconventions.org/Data/cf-standard-names/current/src/cf-standard-name-table.xml is used. Individual standard names are retirved by indexing which return an object of the type CFEntry:

collection = Vocab.CFVocab()
entry = collection["sea_water_temperature"]

source

Base.haskey — Method.

bool = haskey(collection::CFVocab,stdname)

Return true if stdname is part of the NetCDF CF Standard Name vocabulary collection.

source

DIVAnd.Vocab.SDNCollection — Function.

collection = SDNCollection(name)

Open the SeaDataNet collection with the name name at the URL http://www.seadatanet.org/urnurl/collection/ The collection can be indexed with brackets using the identifier.

using DIVAnd
collection = Vocab.SDNCollection("P01")
concept = collection["PSALPR01"]
@show Vocab.prefLabel(concept)

source

DIVAnd.Vocab.prefLabel — Function.

s = Vocab.prefLabel(c::Vocab.Concept)

Return the preferred label of a concept c

source

s = Vocab.prefLabel(urn::AbstractString)

Return the preferred label of a concept usings it URN (Uniform Resource Name)

source

DIVAnd.Vocab.altLabel — Function.

s = Vocab.altLabel(c::Vocab.Concept)

Return the alternative label of a concept c

source

s = Vocab.altLabel(urn::AbstractString)

Return the alternative label of a concept usings it URN (Uniform Resource Name)

source

DIVAnd.Vocab.notation — Function.

s = Vocab.notation(c::Vocab.Concept)

Return the identifier of a concept c

source

s = Vocab.notation(urn::AbstractString)

Return the identifier of a concept usings it URN (Uniform Resource Name)

source

DIVAnd.Vocab.definition — Function.

s = Vocab.definition(c::Vocab.Concept)

Return the definition of a concept c

source

s = Vocab.definition(urn::AbstractString)

Return the definition of a concept usings it URN (Uniform Resource Name)

source

DIVAnd.Vocab.resolve — Function.

entry = Vocab.resolve(urn)

Resolve a SeaDataNet URN (Uniform Resource Name) and returns the corresponding EDMO entry or Vocabulary concept. For example:

concept = Vocab.resolve("SDN:P021:current:TEMP")

source

Base.find — Method.

find(c::Concept,name,collection)

Return a list of related concepts in the collection collection. name can be the string "related", "narrower", "broader".

source

DIVAnd.Vocab.description — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

DIVAnd.Vocab.canonical_units — Function.

str = description(entry::CFEntry)
str = canonical_units(entry::CFEntry)

Return the description or the canonical units of the `entry`.

source

DIVAnd.Vocab.splitURL — Function.

collection,tag,key = Vocab.splitURL(url)

Split a concept URL into collection, tag and key. url must finishe with a slash.

source

Internal API or advanced usage

State vector

DIVAnd.statevector — Type.

Initialize structure for packing and unpacking given their mask.

sv = statevector_init((mask1, mask2, ...))

Initialize structure for packing and unpacking multiple variables given their corresponding land-sea mask.

Input: mask1, mask2,...: land-sea mask for variable 1,2,... Sea grid points correspond to one and land grid points to zero. Every mask can have a different shape.

Output: sv: structure to be used with statevector_pack and statevector_unpack.

Note: see also statevector_pack, statevector_unpack

Author: Alexander Barth, 2009,2017 <a.barth@ulg.ac.be> License: GPL 2 or later

source

DIVAnd.pack — Function.

Pack a series of variables into a vector under the control of a mask.

x = pack(sv,(var1, var2, ...))

Pack the different variables var1, var2, ... into the vector x where sv is a statevector. Only sea grid points are retained.

Input: sv: structure generated by statevector_init. var1, var2,...: variables to pack (with the same shape as the corresponding masks).

Output: x: vector of the packed elements. The size of this vector is the number of elements of all masks equal to 1.

Notes: If var1, var2, ... have an additional trailing dimension, then this dimension is assumed to represent the different ensemble members. In this case x is a matrix and its last dimension is the number of ensemble members.

source

DIVAnd.unpack — Function.

Unpack a vector into different variables under the control of a mask.

var1, var2, ... = unpack(sv,x) var1, var2, ... = unpack(sv,x,fillvalue)

Unpack the vector x into the different variables var1, var2, ... where sv is a statevector.

Input: sv: structure generated by statevector_init. x: vector of the packed elements. The size of this vector is the number of elements equal to 1 in all masks.

Optional input parameter: fillvalue: The value to fill in var1, var2,... where the masks correspond to a land grid point. The default is zero.

Output: var1, var2,...: unpacked variables.

Notes: If x is a matrix, then the second dimension is assumed to represent the different ensemble members. In this case, var1, var2, ... have also an additional trailing dimension.

source

Base.sub2ind — Function.

ind = statevector_sub2ind(sv,subscripts)

Compute from a tuple of subscripts the linear index in the packed state vector. The first element of the subscript indicates the variable index and the remaining the spatial subscripts.

source

Base.ind2sub — Function.

subscripts = ind2ind(sv,index)

Compute from linear index in the packed state vector a tuple of subscripts. The first element of the subscript indicates the variable index and the remaining the spatial subscripts.

source

Base.length — Function.

number of points per node it is always zero for non-leaf nodes

source

Constraints

DIVAnd_constr_fluxes
DIVAnd_constr_constcoast

ODV files

DIVAnd.ODVspreadsheet.listSDNparams — Function.

p = listSDNparam(ODVData)

Return a list of SeaDataNet P01 parameters in a ODV spreadsheet `ODVData`.

source

DIVAnd.ODVspreadsheet.load — Function.

 profiles,lons,lats,depths,times,ids = load(T,fnames,datanames;
    qv_flags = [DIVAnd.ODVspreadsheet.GOOD_VALUE,
                DIVAnd.ODVspreadsheet.PROBABLY_GOOD_VALUE],
    nametype = :P01,
    qvlocalname = "QV:SEADATANET")

Load all profiles in all file from the array fnames corresponding to one of the parameter names datanames. If nametype is :P01 (default), the datanames are P01 vocabulary names with the SDN prefix. If nametype is :localname, then they are the ODV column header without units. For example if the column header is Water body salinity [per mille], then datenames should be ["Water body salinity"]. The resulting vectors have the data type T (expect times and ids which are vectors of DateTime and String respectively). Only values matching the quality flag qv_flags are retained. qv_flags is a vector of Strings (based on http://vocab.nerc.ac.uk/collection/L20/current/, e.g. "1" means "good value"). One can also use the constants these constants (prefixed with DIVAnd.ODVspreadsheet.):

qvlocalname is the column name to denote quality flags. It is assumed that the quality flags follow immediatly the data column.

constant	value

NO_QUALITY_CONTROL	"0"

GOOD_VALUE	"1"

PROBABLY_GOOD_VALUE	"2"

PROBABLY_BAD_VALUE	"3"

BAD_VALUE	"4"

CHANGED_VALUE	"5"

VALUE_BELOW_DETECTION	"6"

VALUE_IN_EXCESS	"7"

INTERPOLATED_VALUE	"8"

MISSING_VALUE	"9"

VALUE_PHENOMENON_UNCERTAIN	"A"

If the ODV does not contain a semantic header (e.g. for the aggregated ODV files), then local names must be used.

julia> data,lon,lat,depth,time,ids = DIVAnd.ODVspreadsheet.load(Float64,["data_from_med_profiles_non-restricted_v2.txt"],
      ["Water body salinity"]; nametype = :localname );

No checks are done if the units are consistent.

source

 profiles,lons,lats,depths,times,ids = load(T,dir,P01names)

Load all ODV files under the directory dir corresponding the one of the parameter names P01names. The resulting vectors have the data type T (expect times and ids which are vectors of DateTime and String respectively).

No checks are done if the units are consistent.

source

DIVAnd.ODVspreadsheet.localnames — Function.

list = localnames(sheet,P01name)

Return a list list of all local names mapping to the specified P01name in the ODV spreadsheet sheet without the prefix "SDN:LOCAL:".

source

list = localnames(sheet)

Return a list list of all local names in the ODV spreadsheet sheet without the prefix "SDN:LOCAL:" in the order as they appear in the ODV file.

source

DIVAnd.ODVspreadsheet.Spreadsheet — Type.

Define composite type that will contain:

the metadata (dictionary),

SDN parameter mapping (dictionary)

the column labels (array) and

the profiles (array of arrays).

source

DIVAnd.ODVspreadsheet.loadprofile — Function.

 data,data_qv,obslon,obslat,obsdepth,obsdepth_qv,obstime,obstime_qv,EDMO,LOCAL_CDI_ID =
 loadprofile(T,sheet,iprofile,dataname; nametype = :P01)

Load a iprofile-th profile from the ODV spreadsheet sheet of the parameter dataname. If nametype is :P01 (default), the dataname is the P01 vocabulary name with the SDN prefix. If nametype is :localname, then it is the ODV column header. The resulting vectors have the data type T (expect the quality flag and obstime) .

source

DIVAnd.ODVspreadsheet.loaddataqv — Function.

data,data_qv = loaddataqv(sheet,profile,locname,fillvalue; fillmode = :repeat)

The same as loaddata, but now the quality flag are also loaded.

profile[i][j] is the j-th column of the i-th row of a profile. profile[i,j] is the i-th column of the j-th row of a profile.

source

DIVAnd.ODVspreadsheet.SDNparse! — Function.

SDNparse!(col,fillmode,fillvalue,data)

Parse the list of String col into the corresponding data type of the vector data. Empty values are either replaced by fillvalue (if fillmode is :fill) or the previous value if repeated (if fillmode is :repeat)

source

DIVAnd.ODVspreadsheet.colnumber — Function.

cn = colnumber(sheet,localname)

Return the column number cn of the first column with the local name localname (without the prefix "SDN:LOCAL:") in the ODV spreadsheet sheet.

source

DIVAnd.ODVspreadsheet.nprofiles — Function.

n = nprofiles(ODVData)

Return the number of profiles in a ODV Spreadsheet ODVData loaded by readODVspreadsheet.

source

Operators

DIVAnd.sparse_interp — Function.

H,out = sparse_interp(mask,I)

Create interpolation matrix from mask and fractional indexes I.

Input: mask: 0 invalid and 1 valid points (n-dimensional array) I: fractional indexes (2-dim array n by mi, where mi is the number of points to interpolate) Ouput: H: sparse matrix with interpolation coefficients out: true if value outside of grid outbbox: 1 if outise bouding box onland: 1 if point touches land (where mask == 0)

source

DIVAnd.sparse_interp_g — Function.

sparse_interp(x,mask,xi) Interpolate from x onto xi

source

DIVAnd.sparse_gradient — Function.

Sparse operator for a gradient. Dx1,Dx2,...,Dxn = sparse_gradient(mask,pmn) Form the gradient using finite differences in all n-dimensions Input: mask: binary mask delimiting the domain. 1 is inside and 0 outside. For oceanographic application, this is the land-sea mask. pmn: scale factor of the grid. Output: Dx1,Dx2,...,Dxn: operators represeting a gradient along different dimensions

source

DIVAnd.sparse_diff — Function.

diffx = sparse_diff(sz1,m,cyclic)

Sparse operator for differentiation along dimension m for "collapsed" matrix of the size sz1. cyclic is true if domain is cyclic along dimension m. false is the default value

source

DIVAnd.matfun_trim — Function.

T = matfun_trim(sz1,m)

Create an operator which trim first and last row (or column) in The field is a "collapsed" matrix of the size sz1. m is the dimension to trim.

source

DIVAnd.matfun_stagger — Function.

S = matfun_stagger(sz1,m,cyclic)

Create an operator for staggering a field in dimension m. The field is a "collapsed" matrix of the size sz1.

Input: sz1: size of rhs m: dimension to stagger cyclic: true if domain is cyclic along dimension m. False is the default value

source

DIVAnd.matfun_diff — Function.

Operator for differentiation.

diffx = matfun_diff(sz1,m,cyclic)

Operator for differentiation along dimension m for "collapsed" matrix of the size sz1.

Input: sz1: size of rhs m: dimension to differentiate cyclic: true if domain is cyclic along dimension m. False is the default value

source

DIVAnd.matfun_shift — Function.

Operator shifting a field in a given dimension.

function S = matfun_shift(sz1,m,cyclic)

Operator shifting a field in the dimension m. The field is a "collapsed" matrix of the size sz1.

Input: sz1: size of rhs m: dimension to shift cyclic: true if domain is cyclic along dimension m. False is the default value

source

Quadtree

DIVAnd.Quadtrees.QT — Type.

quadtree (of the higher-dimensional equivalent) T the type of the coordinates TA the type of the attributes N number of dimensions

source

DIVAnd.Quadtrees.rsplit! — Function.

recursive split

source

DIVAnd.Quadtrees.add! — Function.

sucess = add!(qt,x,attrib,max_cap = 10) Add point x with the attribute attrib to the quadtree qt. sucess is true if xis within the bounds of the quadtree node qt (otherwise false and the point has not been added)

source

DIVAnd.Quadtrees.within — Function.

attribs = within(qt,min,max)

Search all points within a bounding box defined by the vectors min and max.

source

DIVAnd.Quadtrees.bitget — Function.

Test if the n-th bit in a is set. The least significant bit is n = 1.

source

DIVAnd.Quadtrees.inside — Function.

         x1

+–––––+ | | | + | | y | +–––––+ x0

source

DIVAnd.Quadtrees.intersect — Function.

Test of the rectanges defined by x0,x1 and y0,y1 intersects x1 +–––––+ | | | +–––––+ y1 | | | | +–––––+ | x0 | | | | +–––––+ y0

source

DIVAnd.Quadtrees.split! — Function.

split a single node

source

Conjugate gradient

DIVAnd.conjugategradient — Function.

x,success,niter = conjugategradient(fun!,b)

Solve a linear system with the preconditioned conjugated-gradient method: A x = b where A is a symmetric positive defined matrix and b is a vector. Equivalently the solution x minimizes the cost function J(x) = ½ xᵀ A x - bᵀ x.

The function fun!(x,fx) computes fx which is equal to A*x. For example:

function fun!(x,fx)
    fx[:] = A*x
end

Note that the following code will NOT work, because a new array fx would be created and it would not be passed back to the caller.

function fun!(x,fx)
    fx = A*x # bug!
end

The function fun! works in-place to reduce the amount of memory allocations.

Optional input arguments

x0: starting vector for the interations

tol: tolerance on |Ax-b| / |b|

maxit: maximum of interations

pc!: the preconditioner. The functions pc(x,fx) computes fx = M⁻¹ x (the inverse of M times x) where M is a symmetric positive defined matrix. Effectively, the system E⁻¹ A (E⁻¹)ᵀ (E x) = E⁻¹ b is solved for (E x) where E Eᵀ = M. Ideally, M should this be similar to A, so that E⁻¹ A (E⁻¹)ᵀ is close to the identity matrix. The function pc! should be implemented in a similar way than fun! (see above).

Output

x: the solution

success: true if the interation converged (otherwise false)

niter: the number of iterations

source

DIVAnd.pc_none! — Function.

pc_none!(x,fx)

Dummy call-back function when no preconditioner is used. fx will be equal to x.

source

DIVAnd.checksym — Function.

xAy, yATx = checksym(n,fun!)

Check if the the function fun! represents a symmetric matrix when applied on random vectors of size n.

source

Utility functions

DIVAnd.DIVAnd_laplacian
DIVAnd.DIVAnd_obscovar
DIVAnd.DIVAnd_adaptedeps2
DIVAnd.DIVAnd_diagHKobs
DIVAnd.DIVAnd_residual
DIVAnd.DIVAnd_addc
DIVAnd.DIVAnd_erroratdatapoints
DIVAnd.DIVAnd_iBpHtiRHx!
DIVAnd.DIVAnd_GCVKii
DIVAnd.DIVAnd_fittocpu
DIVAnd.DIVAnd_background
DIVAnd.DIVAnd_obs
DIVAnd.DIVAnd_bc_stretch
DIVAnd.DIVAnd_diagHK
DIVAnd.DIVAnd_kernel
DIVAnd.DIVAnd_residualobs
DIVAnd.DIVAnd_aexerr
DIVAnd.DIVAnd_cpme
DIVAnd.DIVAnd_cpme_go
DIVAnd.DIVAnd_datainboundingbox
DIVAnd.DIVAnd_Lpmnrange
DIVAnd.DIVAnd_pc_sqrtiB
DIVAnd.DIVAnd_pc_none
DIVAnd.DIVAnd_GCVKiiobs
DIVAnd.DIVAnd_cutter
DIVAnd.DIVAnd_qc
DIVAnd.DIVAnd_solve!
DIVAnd.DIVAnd_sampler
DIVAnd.DIVAndjog
DIVAnd.DIVAnd_background_components
DIVAnd.stats
DIVAnd.statpos
DIVAnd.blkdiag
Base.findfirst
DIVAnd.formatsize
DIVAnd.interp!
DIVAnd.ufill
DIVAnd.jmBix
DIVAnd.cgradient
DIVAnd.fzero
DIVAnd.localize_separable_grid
DIVAnd.decompB!
DIVAnd.varanalysis
DIVAnd.len_harmonize
DIVAnd.alpha_default
DIVAnd.ncfile
DIVAnd.writeslice
DIVAnd.encodeWMSStyle
DIVAnd.loadoriginators

Examples

To run the example, you need to install PyPlot. In the folder examples of DIVAnd, you can run e.g. the example DIVAnd_simple_example_1D.jl by issuing:

# cd("/path/to/DIVAnd/examples")
include("DIVAnd_simple_example_1D.jl")

Replace /path/to/DIVAnd/ by the installation directory of DIVAnd which is the output of Pkg.dir("DIVAnd") if you installed DIVAnd using Julias package manager.

Performance considerations

Tuning the domain decomposition

The functions diva3d and DIVAndgo split the domain into overlapping subdomains to reduce the required amount of memory. In some circumstances (in particular few vertical levels), this can unnecessarily degrade the performance. The CPU time of the analysis can be improved by increasing the diva3d option memtofit from 3 (default) to higher values (as long as one does not run out of memory). If this parameter is set to a very high value then the domain decomposition is effectively disabled.

Multiple CPU system

Per default julia tries to use all CPUs on your system when doing matrix operations. The number of CPUs is controlled by the call to BLAS.set_num_threads. Using multiple CPUs can result in overhead and it can be beneficial to reduce the number of CPUs:

BLAS.set_num_threads(2)

Information for developers

Update the documentation

Install

Pkg.add("Documenter")

Troubleshooting

If the installation of a package fails, it is recommended to update the local copy of the package list by issuing Pkg.update() to make sure that Julia knows about the latest version of these packages and then to re-try the installation of the problematic package. Julia calls the local copy of the packge list METADATA. For example to retry the installation of EzXML issue the following command:

Pkg.update()
Pkg.add("EzXML")

No plotting window appears

If the following command doesn't produce any figure

using PyPlot
plot(1:10)

A possible solution is to modify the backend: this is done by editing the python configuration file matplotlibrc. The location of this file is obtained in python with:

import matplotlib
matplotlib.matplotlib_fname

Under Linux, this returns '~/.config/matplotlib/matplotlibrc'. To use the TkAgg backend, add the following to the file:

backend      : TkAgg

The matplotlibrc need to be created if it does not exists.

Julia cannot connect to GitHub on Windows 7 and Windows Server 2012

Cloning METADATA or downloading a julia packages fails with:

GitError(Code:ECERTIFICATE, Class:OS, , user cancelled certificate checks: )

The problem is that Windows 7 and Windows Server 2012 uses outdated encryption protocols. The solution is to run the "Easy fix" tool from the Microsoft support page

MbedTLS.jl does not install on Windows 7

The installion of MbedTLS.jl fails with the error message:

INFO: Building MbedTLS                                                                                                                                    
Info: Downloading https://github.com/quinnj/MbedTLSBuilder/releases/download/v0.6/MbedTLS.x86_64-w64-mingw32.tar.gz to C:\Users\Jeremy\.julia\v0.6\MbedTLS
\deps\usr\downloads\MbedTLS.x86_64-w64-mingw32.tar.gz...                                                                                                  
Exception setting "SecurityProtocol": "Cannot convert null to type "System.Net.SecurityProtocolType" due to invalid enumeration values. Specify one of th
e following enumeration values and try again. The possible enumeration values are "Ssl3, Tls"."                                                           
At line:1 char:35                                                                                                                                         
+ [System.Net.ServicePointManager]:: <<<< SecurityProtocol =                                                                                              
    + CategoryInfo          : InvalidOperation: (:) [], RuntimeException                                                                                  
    + FullyQualifiedErrorId : PropertyAssignmentException                                                                                                 
    [...]

See also the issue https://github.com/JuliaWeb/MbedTLS.jl/issues/133

The solution is to install the Windows Management Framework 4.0.

EzXML.jl cannot be installed on RedHat 6

The zlib library of RedHat 6, is slightly older than the library which EzXML.jl and libxml2 requires.

To verify this issue, you can type in Julia

Libdl.dlopen(joinpath(Pkg.dir("EzXML"),"deps/usr/lib/libxml2.so"))

It should not return an error message. On Redhat 6.6, the following error message is returned:

ERROR: could not load library "/home/username/.julia/v0.6/EzXML/deps/usr/lib/libxml2.so"

/lib64/libz.so.1: version `ZLIB_1.2.3.3' not found (required by /home/divahs1/.julia/v0.6/EzXML/deps/usr/lib/libxml2.so)

Stacktrace:

 [1] dlopen(::String, ::UInt32) at ./libdl.jl:97 (repeats 2 times)

However, the following command should work:

 LD_LIBRARY_PATH="$HOME/.julia/v0.6/EzXML/deps/usr/lib/:$LD_LIBRARY_PATH" julia --eval  'print(Libdl.dlopen(joinpath(Pkg.dir("EzXML"),"deps/usr/lib/libxml2.so"))'

Lukily, EzZML.jl includes a newer version of the zlib library, but it does not load the library automatically. (see also https://github.com/JuliaLang/julia/issues/7004 and https://github.com/JuliaIO/HDF5.jl/issues/97)

To make Julia use this library, a user on RedHat 6 should always start Julia with:

LD_LIBRARY_PATH="$HOME/.julia/v0.6/EzXML/deps/usr/lib/:$LD_LIBRARY_PATH" julia

One can also create script with the following content:

#!/bin/bash
export LD_LIBRARY_PATH="$HOME/.julia/v0.6/EzXML/deps/usr/lib/:$LD_LIBRARY_PATH"
exec /path/to/bin/julia "$@"

by replacing /path/to/bin/julia to the full path of your installation directory. The script should be marked executable and it can be included in your Linux search PATH environement variable. Julia can then be started by calling directly this script.

The DIVAnd test suite fails with automatic download failed

Running Pkg.test("DIVAnd") fails with the error:

automatic download failed (error: 2147500036)

The test suite will download some sample data. You need to have internet access and run the test function from a directory with write access.

You can change the directory to your home directory with the julia command cd(homedir()).

You can check the current working directory with:

pwd()

METADATA cannot be updated

Pkg.update fails with the error message METADATA cannot be updated.

If you have git installed, you can issue the command:

cd ~/.julia/v0.6/METADATA
git reset --hard

and then in Julia run Pkg.update() again.

If this does not work, then, you can also delete ~/.julia (https://github.com/JuliaLang/julia/issues/18651#issuecomment-347579521) and in Julia enter Pkg.init(); Pkg.update().

Convert error in DIVAnd_obs

The full error message:

MethodError: Cannot `convert` an object of type DIVAnd.DIVAnd_constrain{Float32,Diagonal{Float64},SparseMatrixCSC{Float64,Int64}} to an object of type DIVAnd.DIVAnd_constrain{Float64,TR,TH} where TH<:(AbstractArray{#s370,2} where #s370<:Number) where TR<:(AbstractArray{#s371,2} where #s371<:Number)
This may have arisen from a call to the constructor DIVAnd.DIVAnd_constrain{Float64,TR,TH} where TH<:(AbstractArray{#s370,2} where #s370<:Number) where TR<:(AbstractArray{#s371,2} where #s371<:Number)(...),
since type constructors fall back to convert methods.

The solution is to use the same type of all input parameters: all Float32 or all Float64.

Monthlist issue

Using comments inside list can lead to unexpected results.

This

 monthlist = [
       [1,2,3]
       #[4,5,6]
       ]

should be written as

 monthlist = [
       [1,2,3]
       ]

Error in the factorisation

The following message

Base.LinAlg.PosDefException(95650)

followed by the stack-trace starting with: julia Stacktrace: [1] #cholfact!#8(::Float64, ::Function, ::Base.SparseArrays.CHOLMOD.Factor{Float64}, ::Base.SparseArrays.CHOLMOD.Sparse{Float64}) at ./sparse/cholmod.jl:1360 ................. [9] DIVAndrun(::BitArray{3}, ::Tuple{Array{Float64,3},Array{Float64,3},Array{Float64,3}}, ::Tuple{Array{Float64,3},Array{Float64,3},Array{Float64,3}}, ::Tuple{Array{Float64,1},Array{Float64,1},Array{Float64,1}}, ::Array{Float64,1}, ::Tuple{Array{Float64,3},Array{Float64,3},Array{Float64,3}}, ::Float64) at /home/ctroupin/.julia/v0.6/DIVAnd/src/DIVAndrun.jl:147

might be due to a wrong choice in the analysis parameters, for example a too long correlation length.